moved activities to proper packages
