+++ brav/app/src/main/java/com/thadocizn/brav/views/PendingCasesActivity.kt
@@ -1,8 +1,9 @@
-package com.thadocizn.brav
+package com.thadocizn.brav.views
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.google.firebase.auth.FirebaseAuth
+import com.thadocizn.brav.R
 import com.thadocizn.brav.models.Case
 import com.thadocizn.brav.services.RetroInstance
 import com.thadocizn.brav.utils.SharedPreference
@@ -10,14 +11,14 @@ import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class CompletedCasesActivity : AppCompatActivity() {
+class PendingCasesActivity : AppCompatActivity() {
 
     private lateinit var idToken: String
     lateinit var sharedPreference: SharedPreference
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_completed_cases)
+        setContentView(R.layout.activity_pending_cases)
 
         sharedPreference = SharedPreference(this)
 
@@ -27,7 +28,7 @@ class CompletedCasesActivity : AppCompatActivity() {
                 if (task.isSuccessful) {
                     idToken = task.result!!.token.toString()
                     println(idToken)
-                    getCompletedCases(idToken)
+                    getPendingCases(idToken)
 
 
                 } else {
@@ -38,10 +39,10 @@ class CompletedCasesActivity : AppCompatActivity() {
             }
     }
 
-    private fun getCompletedCases(token: String) {
+    private fun getPendingCases(token: String) {
 
         val service = RetroInstance().service(token)
-        val call = service.getCompletedCases(sharedPreference.getUserId("userId"))
+        val call = service.getPendingCases(sharedPreference.getUserId("userId"))
 
         call.enqueue(object : Callback<List<Case>> {
             override fun onFailure(call: Call<List<Case>>, t: Throwable) {
