+++ brav/app/src/main/res/layout/activity_completed_cases.xml
@@ -5,6 +5,6 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".CompletedCasesActivity">
+        tools:context=".views.CompletedCasesActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
